Overall changes in application caused by eliminating ref attribute for each quote.
Basically ref was equal to link attr in dom and in Compilation.db so it was eliminated from the whole application.
This means changes everywhere where ref was used as the handle reference to the quote. Everything is now reference as id or link. In the backend (Compilation, quote, etc) link
is called name and id is still id.